updates for migration

+++ src_cuda/gram_id.cu
@@ -105,18 +105,18 @@ inline int32_t diag_piv_dispatcher(cudaStream_t stream, cublasHandle_t handle, d
 
 extern "C" void hyacinXGinterp_bufferSize(int32_t N, int32_t K, hyacinPrecision_t AXtype, hyacinPrecision_t Gtype, uint64_t* dev_work_bytes, uint64_t* pinned_work_bytes) {
   if (N <= 0 || K <= 0) { *dev_work_bytes = *pinned_work_bytes = uint64_t(0); return; }
-  int64_t x_bytes = 0, g_real_bytes = 0;
+  int64_t x_bytes = int64_t(0), g_real_bytes = int64_t(0);
   switch (Gtype) {
-    case HYACIN_F64: { g_real_bytes = sizeof(double); break; } case HYACIN_F32: { g_real_bytes = sizeof(float); break; } 
-    case HYACIN_DD: { g_real_bytes = sizeof(double2); break; } case HYACIN_QF: { g_real_bytes = sizeof(float4); break; } 
-    case HYACIN_F64_COMPLEX: { g_real_bytes = sizeof(double); break; } case HYACIN_F32_COMPLEX: { g_real_bytes = sizeof(float); break; } 
-    case HYACIN_DD_COMPLEX: { g_real_bytes = sizeof(double2); break; } case HYACIN_QF_COMPLEX: { g_real_bytes = sizeof(float4); break; } 
+    case HYACIN_F64: { g_real_bytes = sizeof(double); break; } case HYACIN_F32: { g_real_bytes = sizeof(float); break; }
+    case HYACIN_DD: { g_real_bytes = sizeof(double2); break; } case HYACIN_QF: { g_real_bytes = sizeof(float4); break; }
+    case HYACIN_F64_COMPLEX: { g_real_bytes = sizeof(double); break; } case HYACIN_F32_COMPLEX: { g_real_bytes = sizeof(float); break; }
+    case HYACIN_DD_COMPLEX: { g_real_bytes = sizeof(double2); break; } case HYACIN_QF_COMPLEX: { g_real_bytes = sizeof(float4); break; }
     default: break;
   }
 
   switch (AXtype) {
-    case HYACIN_F64: { x_bytes = sizeof(double); break; } case HYACIN_F32: { x_bytes = sizeof(float); break; } 
-    case HYACIN_F64_COMPLEX: { x_bytes = sizeof(cuDoubleComplex); break; } case HYACIN_F32_COMPLEX: { x_bytes = sizeof(cuComplex); break; } 
+    case HYACIN_F64: { x_bytes = sizeof(double); break; } case HYACIN_F32: { x_bytes = sizeof(float); break; }
+    case HYACIN_F64_COMPLEX: { x_bytes = sizeof(cuDoubleComplex); break; } case HYACIN_F32_COMPLEX: { x_bytes = sizeof(cuComplex); break; }
     default: break;
   }
